+++ data/examples/declaration/class/associated-data1-out.hs
@@ -4,7 +4,6 @@ class Foo a where data FooBar a
 
 -- | Something.
 class Bar a where
-
   -- | Bar bar
   data BarBar a