+++ data/examples/declaration/class/associated-types1-out.hs
@@ -4,7 +4,6 @@ class Foo a where type FooBar a
 
 -- | Something.
 class Bar a where
-
   -- | Bar bar
   type BarBar a