+++ data/examples/declaration/class/newlines-after-where-out.hs
@@ -1,0 +1,5 @@
+class Num a where
+  (+) :: a -> a -> a
+
+class Num a where
+  (+) :: a -> a -> a