+++ data/examples/declaration/class/newlines-and-default-decls.hs
@@ -1,0 +1,13 @@
+class Foo a where
+  foo :: a
+  default foo :: ()
+  foo = ()
+  bar :: a
+  default bar :: ()
+  bar = ()
+
+  qux :: a
+
+  default qux :: ()
+
+  qux = ()