+++ data/examples/declaration/instance/associated-types-out.hs
@@ -3,7 +3,6 @@
 instance Foo Int where type Bar Int = Double
 
 instance Foo Double where
-
   type
     Bar
       Double =