+++ data/examples/declaration/instance/multi-parameter-out.hs
@@ -1,7 +1,5 @@
 instance MonadReader a ((->) a) where ask = id
 
 instance MonadState s (State s) where
-
   get = State.get
-
   put = State.put