+++ data/examples/declaration/instance/newlines-after-where.hs
@@ -1,0 +1,7 @@
+instance Num X where
+  (+) = undefined
+
+instance Num Y where
+
+
+  (+) = undefined